+++ mercury/mercury-library/agent-didcommx/src/main/scala/io/iohk/atala/mercury/AgentService.scala
@@ -6,9 +6,10 @@ import zio._
 import org.didcommx.didcomm.model._
 
 import io.iohk.atala.resolvers._
-import io.iohk.atala.mercury.model.{_, given}
+import io.iohk.atala.mercury.model.{given, _}
 import java.util.Base64
 import io.iohk.atala.mercury.DidComm
+import scala.annotation.nowarn
 
 case class AgentService[A <: Agent](didComm: DIDComm, did: A) extends AgentServiceAny(didComm, did.id)